// Qanat.Web/src/app/shared/generated/model/allocation-plan-minimal-dto.ts
export class AllocationPlanMinimalDto { 
    AllocationPlanID?: number;
    GeographyID?: number;
    GeographyAllocationPlanID?: number;
    ZoneID?: number;
    WaterTypeID?: number;
    LastUpdated?: string;
    WaterTypeName?: string;
    WaterTypeSlug?: string;
    ZoneName?: string;
    ZoneSlug?: string;
    ZoneColor?: string;
    ZoneGroupName?: string;
    ZoneGroupSlug?: string;
    AllocationPeriodsCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}